requirements.txt
asgiref==3.6.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-bootstrap5==23.3
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==19.3.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openpyxl==3.1.2
packaging==23.1
Pillow==9.4.0
pluggy==1.3.0
psycopg2==2.9.6
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.11.1
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.3
typing_extensions==4.6.2
tzdata==2022.7
Werkzeug==2.3.4
